+++ R/ExtendBurnin.R
@@ -15,12 +15,12 @@
 #  Passed to argument \code{cpiter} in \code{\link[hdpx]{hdp_burnin}}
 #  in package hdpx.
 #'
-#' @param verbosity Passed to \code{\link[hdpx]{hdp_posterior}}
-#'      \code{verbosity}.
+#' @param burnin.verbosity Number that controls whether progress
+#'    messages are printed.
 #'
-#' @param hdplist A list representation of
-#'    an \code{\link[hdpx]{hdpState-class}} object;
-#'    a return value \code{\link{Burnin}}.
+#' @param previous.burnin.output Output from
+#'    \code{\link{Burnin}} or the file path of
+#'    a checkpoint file written by \code{\link{Burnin}}.
 #'
 #' @return The same type of object as returned from
 #'  \code{\link{Burnin}.}
@@ -30,21 +30,38 @@
 #' @export
 #'
 ExtendBurnin <-
-  function(hdplist,
-           seedNumber     = 1,
-           burnin    = 4000,
+  function(previous.burnin.output,
+           burnin    = 5000,
            cpiter    = 3,
-           verbosity = 0
+           burnin.verbosity = 0,
+           seedNumber = NULL
   ) {
 
-    as.hdpState <- utils::getFromNamespace(x = "as.hdpState", ns = "hdpx")
+    if(is.character(previous.burnin.output)){
+      if(grepl("mSigHdp.burnin.checkpoint", previous.burnin.output)) {
+        tmp.env <- new.env()
+        load(previous.burnin.output, envir = tmp.env)
+        vars <- ls(tmp.env)
+        if (length(vars) > 1) {
+          stop("The input file does not seem to be a burnin checkpoint")
+        }
+        previous.burnin.output <- tmp.env[[vars]]
 
-    hdp.state <- as.hdpState(hdplist)
+      }else{
+        stop("The input is not a burnin checkpoint")
+      }
+    }
+
+    if(is.null(seedNumber)){
+      seedNumber <- previous.burnin.output$hdplist$seed_activate
+    }
 
     set.seed(seedNumber)
-    output <- hdpx::hdp_burnin (hdp         = hdp.state,
-                                burnin      = burnin,
+
+    output <- hdpx::hdp_burnin (hdp         = previous.burnin.output$hdplist,
+                                burnin      = burnin, # This is the number of burnin iterations
                                 cpiter      = cpiter,
-                                verbosity   = verbosity)
+                                verbosity   = burnin.verbosity)
+    output$likelihood <- c(previous.burnin.output$likelihood, output$likelihood)
     return(invisible(output))
   }